backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==6.0.2
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
